requirements.txt
google-genai
streamlit
python-dotenv
textblob
matplotlib
wordcloud
pandas
langdetect
pycountry
plotly


